cleanup str function testing
