+++ src/a01_way_logic/_test_util/test_a01_str.py
